Adds 'memo' to .copy() ModelMember methods, fixing model copying w/links.
When copying an OpModel whose ModelMember-derived components contain
references to one another, these "links" would break because the members'
.copy() operations did not support any memo dictionary.  This commit adds
memo-dictionary support to all the ModelMember-derived objects, as well as
to the OrderedMemberDict objects that hold them.  Model classes have been
updated to create and utilize a memo dictionary when performing their .copy()
operation, so that now copying models should correctly preserve inter-object
linkages.  Tested on a simple case, but not thoroughly.