Fixes and clarifies "standard" error generator scaling.
The std_error_generators function returns error generators that
are *normalized* so that, as a flattened vector, their frobenius norm == 1.0.
This commit adds comments explaining how this seemingly weird criteria
translates to a dimensional scaling factor.

This commit also upates the std_scale_factor function so it actually
performs the task described in its docstring (previously it returned
the inverse of it!).  This function returns the multiplicative scaling
that, when applied to the output (generators) of std_error_generators,
transforms these generators into those that woud be computed using
non-normalized (Pauli) basis matrices and no prefactors (e.g. no 1/2
in the Hamiltonian generator).  There was also an error in the stochastic
scaling factor for this error generator, which was `d` instead of `d**2`.

This correction to the scaline will alter values reported in the HTML
reports - namely the stochastic and affine projection values - but
should make them consistent with the figure caption.  Extensive comments
have been added to the std_scale_factor explaining how the current
scale values have ben computed.

Finally, this commit removes (comments out) an unused reference to
std_scale_factor in reportables.py.