Updates TreeNode._init_children to default same-class creation.
Previously, when a meta.json file was not present in a child_dir/meta_dir
subdirectory an error would result because this file was needed to obtain
the child class type to create.  This meta.json, however, sometimes may
not exist for good reason, e.g. when a results directory exists but has
never been written as a results directory - just the individual results
have been written.  In this case a ProtocolResultsDir object should still
be able to initialize it's children.  This commit default to creating a
class of the same type as 'self' when a parent is initializing children
and a child directory doesn't contain a meta.json.  It's possible this could
be problematic if the child object *needs* the meta.json for other reasons,
and so just shouldn't be initialized if the meta.json in absent, but I
can't think of a scenario where this happens.