Added qsim and column-stacked bases
QuantumSim is a useful forward simulation package for quantum circuits, and it is often convenient when using that code to work in its native basis. I added this as a ‘qsim’ basis.
Vectorization by column-stacking of density matrices is common, but pygsti favors row-stacking. I added column-stacked matrix units as a ‘col’ basis.