Adds OpModel.circuit_operator method.
A convenience method for computing the operator associated with a
(SPAM-less) Circuit.  This commit also adds a _clean_paramvec guard
on the circuit_layer_operator method so it can be called without
needing to invoke, e.g. model.num_params before it will work (which
seems like a fragile and poor design).  To retain the performance
benefits when this _clean_paramvec call is not needed, the
_circuit_layer_operator method was created and is called internally.