Updates smq2Q_XXII and smq2Q_XXYYII model packs to use layer label convention.
Previously these model packs used gates like "Gxxpi2" to mean parallel
X gates on two qubits.  Since models can hold generic Labels to label
operations, we should be using a [Gx:0Gx:1] type label instead of Gxx:0:1.
This commit performs this update, but keeps the old "condensed" gate label
style available in model packs with a "_condensed" suffix.

A side effect of this is that the target model of a model pack now has
Label objects as the keys of its model.operations dictionary, etc, instead
of simple tuples.  I think it's probably better this way, even if it
results in a more complicated printout of the dictionary.