Limits effect labels in layout to those of POVM in model, even if dataset contains more.
NOTE: this may not be an improvement!  This commit updates the
expand_instruments_and_separate_povm method of Circuit to only return
SeparatePOVMCircuits with effect labels present in the supplied model.
Previously, when observed outcome labels are available (i.e. from a
dataset - the typical case) those labels were taken as the POVM labels
without question.  If the dataset contained other outcomes, e.g. the
counts from leakage states, that were not part of the model, the analysis
code would fail when the model was asked to simulate an effect it didn't
have.  So it seems useful to limit the effects to the effects to those
in the model (what this commit does)
BUT:
1) it does make the code slower (querying the mode for POVM effect labels
 could be a bottleneck, perhaps solveable by caching) AND
2) it may silently ignore circumstances when the user should be alerted
 to the fact that the dataset contains un-modeled outcomes - e.g. all
 the probabilities may not add to 1.0 if there are ignored leakage counts
 included in the dataset.

So... I think this is an improvement, but we may decide to roll it back
in the future.