Updates the way objects are loaded from protocol directories.
In particular, when loading the child subdirectories of an object-type
that has an associated 'meta' subdirectory (e.g. 'edesign' or 'data')
and a child subdirectory is encoutered that does *not* contain the meta
directory, that child is not created.

For example, when a ProtocolData object is loaded from a root directory,
and child subdirectories are uninitialized (have no data/ subdirectory),
then child elements of the parent ProtocolData are not created at loading
time.

This is the desired behavior because, as is the case for ProtocolData,
these children will be created on-demand, and properly initialized, etc,
by the parent object.  Previously the default ProtocolData class type
was used to create empty ProtocolData children and precluding them from
proper on-demand initalization.  Falling back to the parent type is
a fine behavior when we're missing a meta.json, but not when the entire
meta subdirectory is absent -- this indicates the child needs to be constructed.

is loaded from a directory that does *not* contain