Removes the with_times=False directive from multi-pass dataset loading.
After revamping the DataSet-loading code to better distinguish between
a time-dependent DataSet and a time-independent dataset with missing
data (a previous commit), we didn't need to specify with_time=False
to the load_dataset calls in ProtocolData.  However, it seems we only
removed this argument for the case when there is just a single dataset
file, and the with_times=False remained for the case when there were
multiple .txt files to load.  This seems to be just an oversight, as
having different loading behavior based on the number of dataset files
seems confusing to the user.  So this commit removes the with_times=False
argument from the 2+ dataset case.  This enables correct loading of
multiple time-dependent datasets from a data directory (as separate
passes), which was previously broken.