Adds backtracking capability for term stages (path-integral fwdsim).
When beginning a initially optimizing an objective function using
the path-integral technique, we begin at a point where the given set of
paths is "sufficient" - i.e. meets the accuracy requirements demanded
by the user.  This commit adds logic to core.py so that if the paths
are insufficient, and a sufficient set cannot be found (e.g. because of
the maximum paths per circuit) then the model is stepped back toward the
all-zeros parameter vector (the assumption is that this point is easier
to find paths for) until a sufficient path set can be found.

This commit also comments out memory-debugging code with a "# MEM"
prefix in case this is useful in the near future.