Adds more memory efficient way of differentiating & evaluating polynomials.
This is relevant to the path-integral simulation method.  Instead of
differentiating polyomials to create (and store!) the derivatives as
polynomials, only to evaluate the derivatives once, this commit adds
the bulk_eval_compact_polynomials_derivs functions to the opcalc
sub-package.  This function (it's actually multiple since complex/real
cases are treated separately) computes and evaluates the derivates
in one step, never needing to store the derivate polynomials.  This
obviously saves memory, and the savings can be substantial when there
are many qubits involved.

This commit also adds memory-debugging lines, commented out with
"# MEM" prefixes, in termforwardsim.py so that we can easily return
to memory profiling later on.