Adds logic so that the maximum cached term order can be < 1, i.e. == 0.
When memory is in short supply, the term caching performed for the
path-integral forward simulation technique (by objects, i.e. in
operation.py and spamvec.py) can be significant.  In can be desirable to
set MAX_CACHED_TERM_ORDER=0 (the default is 1) to not cache *any*
operation terms.  This however, caused 0-norm Jacobians in multi-qubit
GST tests because it didn't play nicely with always including the
first order terms (so the optimizer could get traction as the
0-paramvec starting point).  This commit fixes this issue and allows
MAX_CACHED_TERM_ORDER=0 and the first order terms to still be included
correctly.  Note that doing this must be performed manually now
(there's no way for users to change MAX_CACHED_TERM_ORDER; it's
hard-coded), and it brings with it a substantial performance penalty.