Adds an optimization for computing many probabilities using CPTP effects.
The evaluation of probabilities when a CPTP POVM is used was running
slowly because the error generator preceding the dot product in
the evaluation of (<<E_i|errorgen)rho>> for many i was evaluated many
times (once for each i).  This commit adds "probability_using_cache"
functions to effect vectors, allowing them to cache a state vector,
e.g. the value of (errorgen|rho>>), to speed up the evaluation of
many probabilities.

This commit also adds "# MEM"-prefixed memory-debugging lines so
that memory profiling can be easily enabled in the future.