Adds a new out-of-bounds checking mode for the LM optimizer.
Previously, the LM optimizer, when performing out-of-bounds checks
(for path-integral forward simulation), would stop whenever an
*attempted* step violated the path-integral error tolerance constraints.
This commit adds (and selects by default, using a hard-coded
oob_check_mode = 1 statement) the ability to halt the optimization only
when a would-be *accepted* step violates the constraints.  In practice
we find this mode to be superior, as it avoids halting the optimization
prematurely at a step that would have been rejected anyway.  There
are reasons for preferring the old, more conservative method, so this
is added as an option (but is currently hard-coded as "on").

This commit also adds commented out memory-checking statements
to the LM optimization.  Lines prefixed with "# MEM" allow this
functionality to easily be turned back on for future memory profiling.