Adds more accurate 'tuned' option for computing the degrees of freedom in a DataSet.
Adds 'tuned' as an available (but not the default) value for the
`method` argument of DataSet.degrees_of_freedom(...).  This option
attempts to account for low-N "Poisson bump" behavior whereby the
expected value of the loglikelihood ratio varies from 0 to ~1.2 to 1.0
as the expected number of outcome counts varies from 0 to 1.5 to infinity.
This many seem largely irrelevant, but for multi-qubit datasets there
can be very small circuit outcome probabilities and thus many circuits
that have only a few observed counts.  The behavior of the 'tuned' mode
is heuristic and still under development.