Adds oob_check_mode as a legitimate LM optimizer option.
Plumbs up the previously hardcoded oob_check_mode option.  In a
prior commit I thought leaving this as hardcoded to `1` (the new
method) was fine, but actually the out-of-bounds functionality is
used in CPTP gauge optimization and must be set to `0` for that
(because a downstream evaluation fails in the `1` case, with a
gauge-transformed map not actually being CPTP).  So now this is a
is an option that can be specified to the optimizer, and the default
is back to the original method (`0`).

In the future, maybe we want to make the default equal `1` for the
path integral forward simulation?