Updates pruned-path system test after recent improvements.
One recent update to the pruned-path (terms) method involved ensuring
that each GST optimization stage (for a given L value) started at a
point (in model parameter space) where a sufficient set of paths could
be found.  If this was impossible given the user-specified limits, the
model's paramvec was contracted toward 0.  The lack of this functionlity
caused our pruned-path system test to find a lower nSigma value than it
should have, given that it was started at such a point and accepted one
iteration (there's a minimum stop iteration check that prevents halting
the optimization on its first outer iteration - maybe that should be
removed?) - so the optimizer got to iteration once for "free" even though
the path set was insufficient.  The resulting model, while it evaluated
the 2DeltaLogL as being low (~4 sigma) really should have been allowed
as the path set was insufficient.  After the update, we find a model
with nSigma ~ 15.

But we also tried changing the heuristic from "scaled" to
"meanscaled" - a less stringent stopping criteria.  This resulted in
finding a solution of ~1 sigma, and we chose to stick with this as the
new system test.