Updates drift report creation; fixes some bugs.
Updates the creation of drift reports so they function again, and
in so doing cleans up some of the argument naming and objects being
passed around.  This commit fixes two bugs where prep_fiducials.index
was used instead of meas_fiducials.index - a bug that only surfaces
when the preparation and measurement fiducials are different.