Adds a new property, Model.num_modeltest_params
This new property defaults to the usual .num_params, but can be set
independently to allow more flexibility in tuning the model tests
within reports and other functions.  This commit moves the logic
involved in testing whether to use the number of non-gauge parameters
instead of the total parameters to the default .num_modeltest_params
property function, simplifying other functions.  It also adds a
memory limit that can be set to avoid computing the number of nongauge
parameters when it would require too much memory.