Adds ability to use compound outcome labels in MultiDataSet text files.
Adds/fixes loading of MultiDataSet text files so that "compound"
outcome labels, i.e. those with intermediate measurement outcomes,
can be specified in the "##Columns " header.  Prior to this commit,
outcome labels such as "0:0:1" would not be parsed as separate
(intermediate) estimates but a single ("0:0:1",) outcome label, which
is probably not what you want.