Continued effort experimenting with wildcard budget optimization.
This commit adds a simple "barrier" optimization method.  It works
better than the simplex method, but takes a long time currently.
More importantly, the simple Newton solver implemented *doesn't*
converge well, indicating the objective funtion optimized at each
t-value (for barrier method) is not nicely behaved.  Adding a
gradient descent fallback of some kind is needed to make the barrier
optimization work.  This indicates we should be looking at how the
objective function could be smoothed out - it seems that the
discontinuities in the Hessian may be causing the problems.

Still a WIP.