Refactors the update_probs method of WildcardBudget.
Now there's a new 'update_circuit_probs' global function in
wildcardbudget.py that can be used independently of a WildcardBudget
object.  This function factors out the core updating of a single
probability distribution given a wildcard budget (e.g. for a single
circuit).  The slow_update_probs method of WildcardBudget utilizes
this as a helper function, the faster update_probs method performs
the same task but in a more optimized way that utilizes caching.