Adds "proxy" wildcard budget aggregate criteria function for testing.
In our ongoing effort to improve the wildcard budget optimization,
this commit tests a smoother "proxy" function for the actual aggregate
log-likelihood constraint on the wildcard budget.  It was expected/hoped
that the discontinuities in the second derivative of the actual function
was what was causing difficulties with the optimization, but initial results
using this proxy function suggest this may not be the case.  When the proxy,
which, at least in the 1Q case being tested mimics the actual function quite
well is used we still see the Newton solver having difficulty converging
quickly to a global optimum.  More testing is required, and this commit
leaves the testing code in gst.py even more of a mess.