Factors experimental wildcard budget optimization code into new wildcard_opt.py
This new module may itself be broken into pieces later on, but this at least
cleans up the gst.py module a bit.  A bunch of work has been done to
eliminate nested functions, and the barrier and cvxopt optimize_*
functions currently work.  The other functions are close to working
but haven't been tested yet and still need to be tweaked/fixed.   The
next step is getting the *_cvxopt_SMALL function working, to see
whether the added SMALL regularization helps cvxopt converge.