Integrates experimental wildcard optimization techniques into API.
Now the bad-fit options contains a `wildcard_methods` member that
controls which wildcard optimization routines (or "chains" of them)
are run to produce one or more wildcard budgets.  Options can be
passed directly to the optimization functions in wildcard_opt.py.
This commit effectively cleans up the wildcard optimization code and
allows the user to specify what optimization algorithms to use
as we continue to explore what works best.  The 'barrier' method
seems promising, the 'neldermead' method is the default and old
method, and the 'cvxopt' method runs but seems to fail for unknown
reasons.  In the future, we should try to understand why the
primal-dual method used by cvxopt doesn't seem to work while the
barrier method does.  For now, this marks a good stopping point
where at least things are cleaned up and accessible to users.