Adds `StateSpaceLabels.nqubits` attribute and sets .dim = inf when appropriate.
When large models are created, we may need to utilize StateSpaceLabels
instances whose dimension is too large for a numpy integer.  This commit
adds protection for this case, so that the .dim attribute gets set to
numpy.inf instead of 0, which is what happends when _np.product silently
overflows.  Since there are a number of operations that just need to know
the number of qubits, this has been added as an attribute of the state
space labels object - for convenience and so that users don't need to
take the log of the full space dimension - which can be infinite.

This allows, e.g., large processor specs containing just a 'clifford'
model to be built, whereas previously they encountered a "taking the
log of 0" error (because the overflowed dimension was 0).