Adds `keep_rows_cols` arguments to Plaquette truncate methods.
This boolean argument allows a parent PlaquetteGridCircuitStructure
object to retain the original number of columns and rows in each of
its plaquettes even after truncating them so that there are entire
rows/cols that are empty.  This is essential for being able to plot
a PlaquetteGridCircuitStructure, as the color box plot plotting
routines assume that each plaquette has the same shape (rows & cols).