Adds ExperimentDesign.truncate_to_design method for further flexibility.
In particular, this functionality is carried to derived classes
so that a minor modification to StandardGSTDesign.copy_with_maxlengths
allows that function to work with incomplete data - i.e. a standard gst
experiment design with circuit lists that don't contain *all* of the
circuits given by the fiducials, germs, etc, can now be copied with
fewer max-lengths.