Adds regularization to wildcard optimization objectives.
Updates the "hard-coded" objective functions like N * log(f / p)
to use the regularized raw objective function objects to that
the wildcard optimization doesn't run amok when probability values
are near (or less than) 0.  This commit also adds better handling for
zero-frequencies to the wildcard optimization routines, which wasn't
included in the initial implementation.  Finally, this commit adds
np.seterr calls to ignore division and invalid errors during
NewtonSolve's backtracking, as it's normal for invalid values to
be encountered when trying to backtrack and the backtracking won't
stop for NaNs (it treats them as not meeting the criteria to
stop backtracking).