First big step to integrating interpygate into pygsti.
Creates InterpolatedDenseOp and InterpolatedOpFactory objects
(within pygsti.extras.interpygate), which derive from DenseOperator
and OpFactory respectively.  These new classes can be instantiated
via the class method `create_by_interpolating_physical_process`
which takes a user-implemented class instance derived from
the new PhysicalProcess class to construct an interpolated gate.

The idea is that pygsti LinearOperator objects should have fast
from_vector methods, whereas PhysicalProcess objects create process
matrices using a more complex physics simulation (and utilize a
`comm` argument).  An InterpolatedDenseOp therefore takes a LinearOperator
as a target operation and a PhysicalProcess as the actual operation,
and interpolates the error generator (which ideally isn't too far
from the identity).  The InterpolatedDenseOp has parameters equal
to those of the PhysicalProcess, potentially preceded by a 'time'
parameter (if the `times` argument is given).  The target operation
can have matching parameters (including the initial time param) to
implement a target that varies with the interpolated parameters.

The InterpolatedOpFactory is similar, though the special 'time'
parameter of the PhyscalProcess can be added as either a factory
argument or as an operator parameter (this needs testing!).  It
seems that both modes could prove useful.

This commit gets the skeleton of this implementation in place and
adds a demo notebook.  Things that still need to be done are:
- test & fix case when physical process has no special 'time'
  parameter.  Maybe it's useful to have a separate physical process
  class for this?
- test case in which InterpolatedOpFactory is base on a time-dependent
  physical process where time is left as an op parameter instead of
  a factory argument.
- add special serialization routines?
- add "aux data" support within physical process (& test it).
- allow `comm` arguments to be None (code assumes a valid comm now)
- polish up API imports, etc.
- add unit tests; make notebook(s) pedegogical.