Adds support for f=0 boundary case when computing wildcard budget.
Adds set "D" within the probability-updating functions in wildcardbudget.py
that holds all the elements for which the frequency (f) == 0.  These
need to be treated specially because the usual probability scaling
(scale f_k by alpha for k s.t. f_k < q_k and beta when f_k > q_k) doesn't
work when f_k == 0.  When there are zero frequencies we want to ignore
them (and set the corresponding p_k = f_k = 0) until they are needed,
since increasing these p_k's doesn't increase the logL as it does for
k values in set A.  Once set A is empty, however, in order to compensate
for decreasing p_k = beta * f_k for k in set B, TVD is "pushed' into
the p_k's for k in set D.  This commit implements this update in
both the slow and fast versions of the probability update code.
Overall, this should make the wildcard budget optimizations more
robust, especially when dealing with low N when f == 0 is more common.
It fixes what would otherwise end in a "TVD should eventually reach 0"
errors.