Adds parameter vector "cleanup" method to Instrument.
A new Instrument._cleanup_paramvec() method syncs the vectors of
(possibly changed) Instrument members with the instrument's saved
parameter vector (self._paramvec).  Previously it was assumed that
users would not change the instrument members directly, but this is
nonintuitive and fragile.  This update allows users to reset the
elements of, e.g. an Instrument member that is a FullDenseOp and
have those updates propagate to the Instrument as they should.