Fixes standard error-generator scaling bugs.
In a prior commit (d7568c9fe14c70408ad47bf77e506551bd39d876) I messed
up the scaling of the stochastic error generators - because the
stochastic_to_linblad function actually had a factor of "d" in it that
was missed.  This commit fixes that issue, and updates the comments
accordingly.

This commit should also fix the error generator "Power" numbers
being wrong because the "standard scaling factor" switched from dividing
to multiplying the projectors (to match the original docstring) in the
prior commit reference above.  We forgot to update this in the computation
of the power in reportables.py, and I expect that fixing this issue
should fix the calcultations of the Power values in reports.