Updates interpolated gate framework to be more general (updates API).
Generalizes the structure of interpolated gate and factory objects.  Now
multiple parameters can be specified by a physical process as "easy to
compute" (though they must be at the end of the parameter ordering),
and interpolated factory objects can now specify which of the interpolated
parameters are the factory arguments (can be a mix of easy/hard-to-compute
interpolated params).  The demo notebook has been updated to
reflect the changes, but more testing is needed - validation that the
MPI distribution is working in particular.  The hope is that this
commit has reached a nearly stable API; the one exception is the
PhysicalProcess class, which might have some methods renamed.