Splits PhysicalProcess class and fixes MPI bugs in interpygate code.
The PhysicalProcess class is split into PhysicalProcess and
PhysicalErrorGenerator classes for clarity.  Some ad-hoc MPI testing
was done and bugs were fixed so that creating interpolated gates and
factories on multiple processors works now (and gives the same result
as when using a single processor).

In an effort to stabilize the API, `do_process_tomography` was renamed
to `run_process_tomography` to be consistent with pygsti naming
conventions.

At this point things are about ready to merge into the main develop
branch, apart from perhaps updating the demo notebook to be more
pedegogical and less focused on testing.