Updates interpolated gate tutorial, changes aux_data => aux_info.
Some final touches on the integration of pygst.extras.interpygate.
This commit revamps the tutorial notebook, making it more informative
and pedegogical, and moves it to the Tutorials/objects/advanced directory.
It also renames "aux_data" to "aux_info" in interpolated gates, as this
is more akin to the aux_info in a DataSet and seems slightly more
appropriate.  Finally, this commit removes the large swatch of commented-out
code from the initial interpygate version that we kept around for
reference but shouldn't need anymore.