Improves memory accounting and handling, specifically for hessian computation.
Another round of improving the memory tracking infrastructure in pygsti,
which uses "array types" to estimate the amount of memory consumed by
a given calculation.  This update fixes some bugs in the memory estimated
for Hessian calculations, and improves the clarity and correctness of
how array types are computed within forward simulation objects.

This commit also plumbs the 'verbosity' argument in several places
where it was left as a stub, so that now Hessian computations properly
output progress when a memory limit is given.