Tests for project_model and more consistent logG-logT generator.
Shifts the old SciPy-based logG-logT to logG-logT-quick in
operation_from_error_generator and changes the base logG-logT to
match the corresponding logic in error_generator.
This change enables turning on the test of logG-logT in
test_err_gen of test_optools.py, but is not sufficient for full
testing of project_model with logG-logT, likely do to approximate
log call on logG in error_generator.