Adds errorgen_coefficients* methods to ComposedOp and EmbeddedOp.
Adds the ability to extract error generator coefficients (as a dict
and arrays) from composed and embedded operators, which allows
errorgen_penalty_factor to be applied to local/cloud-noise models
that contain such operations (only checked on a LocalNoiseModel).

This commit also adds the ErrorMapContainer class, which LindbladSPAMVec
and LindbladPOVM derive from rather than ErrorGeneratorContainer.  This
is needed because ErrorGeneratorContainer must be initialized with an
error generator object and the two classes mentioned above have an
error *map* but this is not necessarily a LindbladOp and so once can't
assume it has a .errorgen attribute.