Fixes instrument diamond norm function so it gives correct results.
We were building the CPTP-map associated with an instrument incorrectly,
by just filling in the diagonal blocks of the augmented map instead of
entire rows.  Initially, we thought this was a basis-related issue, so
this commit also adds some new classical bases 'clgm' and 'clpp', as well
as matrix bases that contain only diagonal (~classical) elements 'clgmmx',
'clppmx', and 'clmx'.  It seems useful to be able to do more with tensor
product bases that include classical parts, and a number of comments have
been added in basis.py to note places where changes may be needed - but
we don't actually implement this functionality yet.

An issue (for future work) is that the Jamiolkowski code needs an operation
with a basis whose dimension is a perfect square.  This works ok when dealing
with direct-sum bases, as these get "embedded" into their larger spaces
correctly.  Tensor-product bases, however, can't do this yet - we probably need
to cleverly implement the create_equivalent and create_simple_equivalent functions
for a tensor product basis - and first figure out what the "standard" ('std')
analogue of a classical bit x qubit (8-dimensional) product basis is... it seems
it should be placed in a qubit x qubit (16-dimensional) space (?).

In any case, this commit should fix the instrument diamond norm issue, and give
some useful starting points for expanding the useability of classical bases.