Removes num_modeltest_params bottleneck and checks gather_slices method.
This commit is likely the first of a number of MPI-related improvements, since
MPI performance seems worse than expected (looking at scaling of up to 64 procs
on a 2Q-GST test case).  This commit swaps num_modeltest_params with simply
num_params during the per-iteration reporting, since the former can be a bottleneck
in 2Q-GST and larger cases.  The remaining bottleneck seems to be the gather_slices
method in _bulk_fill_dprobs (in distforwardsim.py).  We thought this might be
running slow because it calls comm.Bcast repeatedly instead of a call like Allgatherv,
but this turned out to be incorrect - gather_slices (in mpitools.py) now has a
commented out block, marked for removal, that implements an Allgatherv approach for
the special case that it's applicable, but this didn't actually speed anything up
significantly.