Adds ability to use PrimitiveOpsWildcardBudget with error-rate (opless) models.
Adapts ErrorRatesModel so it is compatible with a PrimitiveOpsWildcardBudget.
This is accomplished by adding primitive_op_labels and primitive_instrument_labels
properties to ErrorRatesModel objects and adding support for an 'idle_name'
to PrimitiveOpsWildcardBudget.

Additionally, this commit expands the idle_name functionality in ErrorRatesModel
and PrimitiveOpsWildcardBudget so that idle_name can be set to 'None'.  In this
case, circuit layers are processed without padding with idles, and in either
case they are processed by components recursively.

Also, in an ErrorRatesModel, op/gate labels can be referenced by name when a full
match is not found (e.g. label 'Gx' works for 'Gx:0' and 'Gx:1').

Other minor bugs fixed in wildcard budget workflow.