Updates jacobian computation to use shared memory and lessen interprocessor communication.
This is an initial and experimental update, whereby jacobian memory within
an objective function object (a "store") is allocated using memory shared
between processors on the same node/host (distinguished by hostname currently).
The ResoureAllocation object now has "host_comm" and "interhost_comm" attributes,
which can be used (when not None) to both indicate that shared memory is being
used and provide the needed communicators for efficiently gathering results.

In particular, our strategy is as follows:
- each comm is replaced with a ResourceAllocation object, which, in addition
 to the comm contains several sub-comms (comms containing all or a subset of
 the original comm's processors).
  - a "host_comm", which contains just the other processors on the same host.
  - an "interhost_comm", which contains one processor per host.
- we assume that the number of processors per host is always constant (even
 after splitting a comm, which may lead to an assertion fail if this isn't
 true), so that all the "interhost" comms are disjoint (not overlapping).
- when gathering data (so far just in gather_slices, which may be all
 we use/need), the interhost comms are used to broadcast data: a list of
 the blocks to be broadcast are iterated over, and when a processor
 encounters a block whose owner is not within its interhost comm it
 simply skips the block, knowing that the owner *is* a part of another
 interhost comm and those processors will distribute the data across
 all of the hosts.

This functionality has been lightly tested in 2Q cases where processor
distribution is performed column-wise (1 atom and many param blocks).
Also, so far the shared memory only applies to the jacobian memory -
in the future we could extend this to the function vector itself and
possibly the Hessian (?).