Minor fix to shared-memory update and an attempt to fix resource_tracker issues.
Garbage collection of shared memory objects appears to be a well-documented
pain.  When MPI processes exit after calling shared_mem.close() they also
call .unlink() via the resource_manager that automatically registers all
shared memory.  This results in tons of warning from processors that try
to close or unlink the shared memory after another process has already
done it (apparently the situation is nicer when all the processes sharing
the memory have a the same parent process, but that isn't the case for us).