Adds debug statements and fixes resource_tracker workaround.
Now the multiprocessing.resource_tracker workaround (that avoids
annoying warnings about leaded shared memory) actually works -- we
forgot to prepend a '/' to the shared memory name before calling
resource_tracker.unregister.  This commit also adds a number of
debug statement that are commented out and will be removed in the
next commit.