Upgrades heuristic for number of atoms in MapForwardSimulator.create_layout
Previously MapForwardSimulator would just divide the computation into
nAtoms = nProcessors parts.  This isn't too bad, but becomes inefficient
when the task portions assigned to each processor are overwhelmed by the
cost of calling from_vector on the model.  When this happens, it would
be better to create fewer atoms (divisions of circuits) and more divisions
in the parameter dimension, as this reduces the number of from_vector calls
needed in the finite-difference jacobian computation.  To this end, this
commit adds a heuristic whereby MapForwardSimulator.create_layout initially
tries to create a layout with nAtoms = 2 * model.dim when this is less than
the number of processors.  This is just a heuristic, and can probably be
improved upon.  This commit also adds a num_atoms argument to
MapForwardSimulator.__init__ so the user can specify the number of atoms
directly if desired.