Adds FOGI setup and initial functionality to ExplicitOpModel.
Maybe this implementation is more general and should be moved to a
base class later on, but for now we add the ability to get & set
the coefficients of first-order-gauge-invariant (FOGI) error generators
(linear combos of elementary error gens) via the `fogi_errorgen_coefficients_array`
and `fogi_errorgen_coefficient_labels` methods.  The FOGI generators
are found using Robin's nullspace-computing method, called within the
fogi_setup method, which must be called before the other methods will work.

There's also preliminary, completely untested, support to add an
interposer that causes an explicit model's parameter to be those
FOGI quantities, but this requires that the operations be parameterized
using a non-positive parameterization, e.g. H+s or GLND.