Debugs the nullspace-based FOGI-finding method for reduced models & other bug fixes.
Replaces the ad-hoc (and incorrect) initial method of zeroing-out any
FOGI column vector that had weight on a row corresponding to an elementary
error generator not present in the model.  The new method first computes
the gauge space of the reduced model (see
ExplicitOpModel._construct_gauge_space_for_model) and then utilizes
that throughout the remainder of the calculation.  (Currently, for
historic reasons, we use np.delete to remove rows from each
op-set's gauge action matrix that is inefficient.  We could just
construct the FOGI action matrix using the *present* elem error gens
in the FUTURE.)

This commit also fixes a bug in the ordering of the FOGI coefficients
when `complete=True` so that they properly match the labels (H & S gauge
at the *end*, not just following their respective FOGI directions).

There's still a number of debug statements commented out throughtout,
and the code is inefficient but hopefully clear.  Could use some cleaup
later.