Adds "constructive" approach to finding FOGI quantities.
This new approach finds FOGI quantities and a reference frame
(a space upon which the gauge acts faithfully) for individual gates,
then combines groups of gates to form larger groups, "juicing" out
FOGI quantities as it proceeds.  The algorithm is still crude, and
could be made more efficient by keeping track of what "dangling"
reference frame directions a gate still needs to "match" with another
gate/group, so that merging pairs of groups is not considered
needlessly.  But the current implementation does work, and has been
successfully compared with the nullspace method for 1Q, 2Q, and a
reduced-2Q model.