Adds a new circular layout for displaying FOGI quantities.
This new & improved (I think) way of displaying FOGI quantities places
the gates along the perimeter of a circle and groups them by their
target qubits.  It seems a bit nicer than the initial attempt, as it
makes more room for edges (it was motivated in part by the need to
display a near-fully connected graph).

The new diagram is created using the render_circle method, and the
old "render" method has become render_grid.